services/api-gateway/pyproject.toml
[project]
name = "eve-api-gateway"
version = "1.0.0"
description = "EVE Co-Pilot API Gateway - Routes requests to microservices"
requires-python = ">=3.11"
dependencies = [
    "eve-shared",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyJWT>=2.8.0",
    "PyYAML>=6.0",
    "redis>=5.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
